common summary method
